# tests_docs/api_contract.yaml
openapi: 3.0.3
info:
  title: RideSmart API
  version: "0.1.0"
  description: >
    AI-Based Smart Public Transportation Platform.
    Real-time bus tracking, ETA prediction and crowd-aware route planning.

    OWNER: Member 2 (database/ - FastAPI application).
    CONTRACT OWNER: Member 5 (tests_docs/). Treat this file as the source of
    truth. If the code and this file disagree, that is a bug.
servers:
  - url: http://localhost:8000
    description: Local development
...
paths:
...
  /api/routes/{route_id}/stops:
    get:
      tags: [routes]
      summary: Ordered stops for a route
      parameters:
        - name: route_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Stops in travel order
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/RouteStop" }
        "404":
          description: Route not found
  /api/routes/{route_id}/shape:
    get:
      tags: [routes]
      summary: Real road geometry for a route
      description: >
        The route's OSRM road polyline in travel order. The map draws this and
        the simulator walks it, so the drawn line is the road the bus is on.
        Stop coordinates alone are not a route - adjacent stops can be
        kilometres apart with a lake or a park between them.
      parameters:
        - name: route_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Polyline vertices in travel order
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RouteShape" }
        "404":
          description: Route not found
...
  /api/buses/active:
    get:
      tags: [tracking]
      summary: Latest known position of every active bus
      description: >
        Returns one entry per active trip. Frontend polls this every 2 seconds
        as a fallback if the WebSocket is unavailable.

        `route_id` is optional. Omitting it returns the whole fleet, which is
        what the fleet-wide panels use. Passing it narrows the response to one
        route, which is what the live map's polling fallback uses.
      parameters:
        - name: route_id
          in: query
          required: false
          schema: { type: integer }
      responses:
        "200":
          description: Active buses
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/BusPosition" }
  /api/buses/{bus_id}/location:
    get:
      tags: [tracking]
      summary: Latest position of one bus
      parameters:
        - name: bus_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Latest position
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BusPosition" }
        "404":
          description: Bus not found
...
components:
  schemas:
    Stop:
      type: object
      required: [id, code, name, lat, lon]
      properties:
        id: { type: integer, example: 1 }
        code: { type: string, example: STOP_VIT }
        name: { type: string, example: VIT }
        lat: { type: number, format: double, example: 12.968142 }
        lon: { type: number, format: double, example: 79.156252 }
        accessible: { type: boolean, example: true }
        kind:
          type: string
          enum: [transit, campus]
          default: transit
          description: >-
            'transit' = OSM maps a real bus stop / bus station / station here.
            'campus' = anchored at a landmark because no bay was surveyed
            there, which is the case for Kingston Engineering College.
...
    ShapePoint:
      type: object
      required: [seq, lat, lon, cum_m]
      properties:
        seq: { type: integer, example: 0 }
        lat: { type: number, format: double, example: 12.968078 }
        lon: { type: number, format: double, example: 79.156234 }
        cum_m: { type: number, format: double, example: 0.0 }
    RouteShape:
      type: object
      required: [route_id, code, total_m, point_count, points]
      properties:
        route_id: { type: integer, example: 1 }
        code: { type: string, example: "V1" }
        total_m: { type: number, format: double, example: 15157.7 }
        point_count: { type: integer, example: 92 }
        points:
          type: array
          items: { $ref: "#/components/schemas/ShapePoint" }
...
    RouteStop:
      type: object
      required: [seq, stop]
      properties:
        seq: { type: integer, description: Zero-based position in travel order }
        stop: { $ref: "#/components/schemas/Stop" }
        scheduled_offset_sec:
          type: integer
          description: Seconds from trip start to scheduled arrival here
    BusPosition:
      type: object
      required: [bus_id, lat, lon, ts]
      properties:
        bus_id: { type: integer, example: 1 }
        bus_reg: { type: string, example: TN09AB1234 }
        trip_id: { type: integer, example: 1 }
        route_id: { type: integer, example: 21 }
        route_code: { type: string, example: "V1" }
        lat: { type: number, format: double, example: 12.968078 }
        lon: { type: number, format: double, example: 79.156234 }
        heading: { type: number, description: Degrees from north }
        speed_kmph: { type: number, example: 28.4 }
        next_stop_name: { type: string, example: Katpadi Junction- Chittor Bus Stand }
        crowd_level: { type: string, enum: [low, med, high] }
        crowd_load: { type: integer, example: 35 }
        capacity: { type: integer, example: 50 }
        wheelchair_accessible: { type: boolean, example: true }
        low_floor: { type: boolean, example: true }
        ts: { type: string, format: date-time }
...